+++ spec/spec_helper.rb
@@ -23,4 +23,8 @@ RSpec.configure do |config|
   config.expect_with(:rspec) do |c|
     c.syntax = :expect
   end
+
+  config.before(:each) do
+    EmailInquire.custom_invalid_domains = nil
+  end
 end